Add option to disable furrr and use mclapply in runner
Introduces the 'bregr.use_furrr' option to allow disabling furrr-based parallelism in the runner function. When set to FALSE, the code falls back to using parallel::mclapply, with a warning that this does not work on Windows. This provides more flexibility for users who may not want to use furrr.